Add better sql error handling
Sql queries for db and user mng will run in a DB transaction.
Running multiple queires separated by a semi-colon will only return an
error for the first query.
Event though transactions are not always used for data consistency it
is necessary to keep the context after a 'using' statement